Add files via upload
Reproducible Research Week 2 Project 1: Note I could not get the .md file to plot the graphs.  I reviewed a lot of comments on google and nothing I did worked.  I did not have time (given going out of town) and was hesitant to uninstall and reinstall R & R Studio.  The plots are contained in the "unnamed-chunk-2-1.png etc.